Break up MotorController into a specialized subclass for Phoenix Motor Controllers.
