+++ core/src/main/kotlin/org/sert2521/sertain/motors/MotorPidf.kt
@@ -22,7 +22,7 @@ class MotorPidfConfig : PidfConfig() {
     var period: Int? = null
 }
 
-class MotorPidfCollection(val motor: MotorController, vararg initialPidf: Pair<Int, MotorPidf>) {
+class MotorPidfCollection(val motor: PhoenixMotorController, vararg initialPidf: Pair<Int, MotorPidf>) {
     private val pidfMap = initialPidf.toMap(mutableMapOf())
 
     fun toMap() = pidfMap
